typing api-actions and refactor type
