+++ packages/@coorpacademy-app-review/src/actions/api/fetch-slide.ts
@@ -1,5 +1,6 @@
+import type {Action, Dispatch} from 'redux';
 import buildTask from '@coorpacademy/redux-task';
-import {Options, SlideFromAPI} from '../../types/common';
+import {Options, SlideFromAPI, StoreState} from '../../types/common';
 
 export const SLIDE_FETCH_REQUEST = '@@slides/FETCH_REQUEST' as const;
 export const SLIDE_FETCH_SUCCESS = '@@slides/FETCH_SUCCESS' as const;
@@ -10,8 +11,9 @@ export type ReceivedSlide = {
   payload: SlideFromAPI;
 };
 
-export const fetchSlide = (slideRef: string, token: string) =>
+export const fetchSlide = (slideRef: string, token: string): Action =>
   buildTask({
     types: [SLIDE_FETCH_REQUEST, SLIDE_FETCH_SUCCESS, SLIDE_FETCH_FAILURE],
-    task: (dispatch, getState, {services}: Options) => services.fetchSlide(slideRef, token)
+    task: (dispatch: Dispatch, getState: () => StoreState, {services}: Options) =>
+      services.fetchSlide(slideRef, token)
   });
