+++ packages/@coorpacademy-app-review/src/actions/api/post-progression.ts
@@ -1,5 +1,6 @@
+import type {Action, Dispatch} from 'redux';
 import buildTask from '@coorpacademy/redux-task';
-import {Options, ProgressionFromAPI} from '../../types/common';
+import {Options, ProgressionFromAPI, StoreState} from '../../types/common';
 import {fetchSlide} from './fetch-slide';
 
 export const POST_PROGRESSION_REQUEST = '@@progression/POST_REQUEST' as const;
@@ -14,10 +15,10 @@ export type ReceiveProgression = {
   };
 };
 
-export const postProgression = (skillRef: string, token: string) =>
+export const postProgression = (skillRef: string, token: string): Action =>
   buildTask({
     types: [POST_PROGRESSION_REQUEST, POST_PROGRESSION_SUCCESS, POST_PROGRESSION_FAILURE],
-    task: (dispatch, getState, {services}: Options) => {
+    task: (dispatch: Dispatch, getState: () => StoreState, {services}: Options) => {
       return services.postProgression(skillRef, token).then((progression: ProgressionFromAPI) => {
         const {ref} = progression.state.nextContent;
         return dispatch(fetchSlide(ref, token));
