+++ packages/@coorpacademy-app-review/src/actions/ui/step-items.ts
@@ -1,19 +1,10 @@
+import {IconValue} from '../../types/common';
 import {FinishedSlidesState} from '../../reducers/ui/finished-slides';
 
 // -----------------------------------------------------------------------------
 
 export const UPDATE_STEP_ITEMS_ON_VALIDATION = '@@step-items/UPDATE_STEP_ITEMS_ON_VALIDATION';
 export const UPDATE_STEP_ITEMS_ON_NEXT = '@@step-items/UPDATE_STEP_ITEMS_ON_NEXT';
-
-const ICON_VALUES = {
-  right: 'right',
-  wrong: 'wrong',
-  'no-answer': 'no-answer'
-} as const;
-
-// -----------------------------------------------------------------------------
-
-export type IconValue = keyof typeof ICON_VALUES;
 
 type UpdateStepItemsOnValidationClickPayload = {
   stepNumber: number;
