+++ packages/@coorpacademy-app-review/src/reducers/data/token.ts
@@ -1,17 +1,8 @@
+import {TokenState} from '../../types/common';
 import {StoreToken, STORE_TOKEN} from '../../actions/data/token';
 
-// -----------------------------------------------------------------------------
-
-export type TokenState = string | null;
-
-// -----------------------------------------------------------------------------
-
-export const initialState: TokenState = null;
-
-// -----------------------------------------------------------------------------
-
 // eslint-disable-next-line default-param-last
-const reducer = (state: TokenState = initialState, action: StoreToken): TokenState => {
+const reducer = (state: TokenState = null, action: StoreToken): TokenState => {
   switch (action.type) {
     case STORE_TOKEN:
       return action.payload;
@@ -19,7 +10,5 @@ const reducer = (state: TokenState = initialState, action: StoreToken): TokenSta
       return state;
   }
 };
-
-// -----------------------------------------------------------------------------
 
 export default reducer;
